feat: unify Dioxus components into frontend crate
