+++ server/Cargo.toml
@@ -14,6 +14,7 @@ dioxus-ssr = "0.7"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 omnibus-shared = { path = "../shared" }
+omnibus-frontend = { path = "../frontend", features = ["web"] }
 sqlx = { version = "0.8", features = ["runtime-tokio-rustls", "sqlite", "macros"] }
 tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
