--- server/src/frontend/components/mod.rs
@@ -1,2 +1,0 @@
-pub mod nav;
-pub use nav::TopNav;
