--- server/src/frontend/components/nav.rs
@@ -1,11 +1,0 @@
-use dioxus::prelude::*;
-
-#[component]
-pub fn TopNav() -> Element {
-    rsx! {
-        nav { class: "top-nav",
-            a { href: "/", "Home" }
-            a { href: "/settings", "Settings" }
-        }
-    }
-}
