--- server/src/frontend/pages/mod.rs
@@ -1,5 +1,0 @@
-pub mod landing;
-pub mod settings;
-
-pub use landing::LandingPage;
-pub use settings::SettingsPage;
